refactor: 로또 가격을 BigInteger 타입으로 변경하라

+++ src/main/java/lotto/domain/LottoConstants.java
@@ -1,9 +1,11 @@
 package lotto.domain;
 
+import java.math.BigInteger;
+
 public class LottoConstants {
     public static final int COUNT_OF_LOTTO_NUMBERS = 6;
     public static final int MIN_LOTTO_NUMBER = 1;
     public static final int MAX_LOTTO_NUMBER = 45;
 
-    public static final int LOTTO_PRICE = 1000;
+    public static final BigInteger LOTTO_PRICE = BigInteger.valueOf(1000);
 }
