+++ src/test/java/lotto/domain/LottoSellerTest.java
@@ -76,7 +76,7 @@ class LottoSellerTest {
     @DisplayName("로또 1개 가격과 구입 금액을 기준으로 구매한 수량과 크기가 같은 LottoTicket을 생성한다.")
     @ParameterizedTest
     @CsvSource({"5000,0,0", "10000,70000,7", "1,8000,8000", "30000,360000,12", "100,95000000,950000"})
-    void createLottoTicketBy(int lottoPrice, BigInteger purchaseAmount, int expectedQuantity) {
+    void createLottoTicketBy(BigInteger lottoPrice, BigInteger purchaseAmount, int expectedQuantity) {
         LottoMachine lottoMachine = LottoMachineFixture.createLottoMachine();
         LottoSeller lottoSeller = new LottoSeller(lottoPrice, lottoMachine);
